Make graph copying in karateclub_node2vec_train explicit.
When reading the code for karateclub_node2vec_train and running across nx.relabel_nodes, it might not be clear that nx.relabel_nodes returns a copy since copy=True is a default parameter value. This commit makes the code read more explicitly. This commit is intended to be a no-op for karateclub_node2vec_train.